refact: directory structure
