+++ crates/binding/src/lib.rs
@@ -9,8 +9,8 @@ use js_plugin::JsPlugin;
 use mako::compiler::{Args, Compiler};
 use mako::config::Config;
 use mako::dev::DevServer;
-use mako::logger::init_logger;
 use mako::plugin::Plugin;
+use mako::utils::logger::init_logger;
 use napi::bindgen_prelude::*;
 use napi::{JsObject, Status};
 use tsfn::{JsHooks, TsFnHooks};
