+++ crates/mako/src/ast/css_ast.rs
@@ -13,13 +13,12 @@ use mako_core::swc_ecma_parser::StringInput;
 use mako_core::{md5, swc_atoms, swc_css_parser, swc_css_visit};
 use swc_core::common::FileName;
 
-use super::file::Content;
-use crate::ast::file::File;
+use crate::ast::file::{Content, File};
+use crate::ast::sourcemap::build_source_map_to_buf;
 use crate::ast::{error, utils};
 use crate::compiler::Context;
 use crate::config::{DevtoolConfig, Mode};
 use crate::module::Dependency;
-use crate::sourcemap::build_source_map_to_buf;
 use crate::visitors::css_dep_analyzer::CSSDepAnalyzer;
 
 #[derive(Clone)]
