+++ crates/mako/src/ast/mod.rs
@@ -1,7 +1,9 @@
+pub(crate) mod comments;
 pub(crate) mod css_ast;
 pub(crate) mod error;
 pub mod file;
 pub(crate) mod js_ast;
+pub(crate) mod sourcemap;
 #[cfg(test)]
 pub mod tests;
 pub(crate) mod utils;
