+++ crates/mako/src/generate/hmr.rs
@@ -6,9 +6,9 @@ use mako_core::swc_ecma_ast::{
 };
 
 use crate::ast::js_ast::{JSAstGenerated, JsAst};
-use crate::chunk::Chunk;
 use crate::compiler::Compiler;
-use crate::generate_chunks::modules_to_js_stmts;
+use crate::generate::chunk::Chunk;
+use crate::generate::generate_chunks::modules_to_js_stmts;
 use crate::module::ModuleId;
 
 impl Compiler {
@@ -21,7 +21,7 @@ impl Compiler {
     ) -> Result<(String, String)> {
         let module_graph = &self.context.module_graph.read().unwrap();
         let (js_stmts, _) = modules_to_js_stmts(module_ids, module_graph, &self.context).unwrap();
-        let mut content = include_str!("runtime/runtime_hmr.js").to_string();
+        let mut content = include_str!("../runtime/runtime_hmr.js").to_string();
         content = content.replace("__CHUNK_ID__", &chunk.id.id).replace(
             "__runtime_code__",
             &format!("runtime._h='{}';", current_hash),
