+++ crates/mako/src/generate/mod.rs
@@ -1,3 +1,15 @@
+pub(crate) mod chunk;
+pub(crate) mod chunk_graph;
+pub(crate) mod chunk_pot;
+pub(crate) mod generate_chunks;
+pub(crate) mod group_chunk;
+pub(crate) mod hmr;
+pub(crate) mod minify;
+pub(crate) mod optimize_chunk;
+pub(crate) mod runtime;
+pub(crate) mod swc_helpers;
+pub(crate) mod transform;
+
 use std::collections::{HashMap, HashSet};
 use std::fs;
 use std::ops::DerefMut;
@@ -13,11 +25,11 @@ use mako_core::tracing::debug;
 
 use crate::compiler::{Compiler, Context};
 use crate::config::{DevtoolConfig, OutputMode, TreeShakingStrategy};
-use crate::generate_chunks::{ChunkFile, ChunkFileType};
+use crate::dev::update::UpdateResult;
+use crate::generate::generate_chunks::{ChunkFile, ChunkFileType};
 use crate::module::{Dependency, ModuleId};
 use crate::stats::{create_stats_info, print_stats, write_stats};
-use crate::update::UpdateResult;
-use crate::util::base64_encode;
+use crate::utils::base64_encode;
 use crate::visitors::async_module::mark_async;
 
 #[derive(Clone)]
