+++ crates/mako/src/generate/swc_helpers.rs
