+++ crates/mako/src/plugin.rs
@@ -8,9 +8,9 @@ use mako_core::swc_common::Mark;
 use mako_core::swc_ecma_ast::Module;
 
 use crate::ast::file::{Content, File};
-use crate::chunk_graph::ChunkGraph;
 use crate::compiler::{Args, Context};
 use crate::config::Config;
+use crate::generate::chunk_graph::ChunkGraph;
 use crate::module::{Dependency, ModuleAst};
 use crate::module_graph::ModuleGraph;
 use crate::stats::StatsJsonMap;
