+++ crates/mako/src/plugins/minifish.rs
@@ -14,8 +14,8 @@ use serde::Serialize;
 use unsimplify::UnSimplify;
 
 use crate::ast::file::{Asset, Content, JsContent};
+use crate::build::load::FileSystem;
 use crate::compiler::Context;
-use crate::load::FileSystem;
 use crate::module::{Dependency as ModuleDependency, ModuleAst, ResolveType};
 use crate::plugin::{Plugin, PluginLoadParam, PluginParseParam, PluginTransformJsParam};
 use crate::plugins::bundless_compiler::to_dist_path;
