+++ crates/mako/src/stats.rs
@@ -13,9 +13,9 @@ use mako_core::pathdiff::diff_paths;
 use mako_core::serde::Serialize;
 use swc_core::common::source_map::Pos;
 
-use crate::chunk::ChunkType;
 use crate::compiler::Compiler;
 use crate::features::rsc::{RscClientInfo, RscCssModules};
+use crate::generate::chunk::ChunkType;
 
 #[derive(Debug, PartialEq, Eq)]
 // name 记录实际 filename , 用在 stats.json 中, hashname 用在产物描述和 manifest 中
