+++ crates/mako/src/utils/mod.rs
@@ -1,3 +1,10 @@
+pub mod logger;
+pub mod profile_gui;
+#[cfg(test)]
+pub(crate) mod test_helper;
+pub(crate) mod thread_pool;
+pub mod tokio_runtime;
+
 use mako_core::anyhow::{anyhow, Result};
 use mako_core::base64::engine::general_purpose;
 use mako_core::base64::Engine;
